+++ backend/canisters/user/impl/src/openchat_bot.rs
@@ -1,10 +1,11 @@
+use crate::RuntimeState;
 use crate::updates::c2c_send_messages::{HandleMessageArgs, handle_message_impl};
-use crate::{Membership, RuntimeState};
 use chat_events::{MessageContentInternal, TextContentInternal};
 use constants::{DAY_IN_MS, HOUR_IN_MS, OPENCHAT_BOT_USER_ID, OPENCHAT_BOT_USERNAME};
 use types::nns::Tokens;
 use types::{ChannelId, CommunityId, EventWrapper, Message, SuspensionDuration, User, UserId, UserType};
 use user_canister::{C2CReplyContext, PhoneNumberConfirmed, StorageUpgraded, UserSuspended};
+use user_state::Membership;
 use utils::format::format_to_decimal_places;
 
 pub(crate) fn send_community_deleted_message(deleted_by: UserId, name: String, public: bool, state: &mut RuntimeState) {